+++ addon/inspect.js
@@ -371,13 +371,14 @@ class Model {
     }
 
     // Fetch fields using a Tooling API call, which returns fields not readable by the current user, but fails if the user does not have access to the Tooling API.
-    // The Tooling API is not very stable. It often gives "An unexpected error occurred. Please include this ErrorId if you contact support".
     // We would like to query all meta-fields, to show them when the user clicks a field for more details.
     // But, the more meta-fields we query, the more likely the query is to fail, and the meta-fields that cause failure vary depending on the entity we query, the org we are in, and the current Salesforce release.
-    // Therefore qe query the minimum set of meta-fields needed by our main UI.
+    // Therefore we query the minimum set of meta-fields needed by our main UI.
     this.spinFor(
       "querying tooling particles",
-      sfConn.rest("/services/data/v" + apiVersion + "/tooling/query/?q=" + encodeURIComponent("SELECT QualifiedApiName, Label, DataType, ReferenceTo, Length, Precision, Scale, IsAutonumber, IsCaseSensitive, IsDependentPicklist, IsEncrypted, IsIdLookup, IsHtmlFormatted, IsNillable, IsUnique, IsCalculated, InlineHelpText, FieldDefinition.DurableId, EntityDefinition.DurableId FROM EntityParticle WHERE EntityDefinition.QualifiedApiName = '" + this.sobjectName + "'")).then(res => {
+      //sfConn.rest("/services/data/v" + apiVersion + "/tooling/query/?q=" + encodeURIComponent("SELECT QualifiedApiName, Label, DataType, ReferenceTo, Length, Precision, Scale, IsAutonumber, IsCaseSensitive, IsDependentPicklist, IsEncrypted, IsIdLookup, IsHtmlFormatted, IsNillable, IsUnique, IsCalculated, InlineHelpText, FieldDefinition.DurableId, EntityDefinition.DurableId FROM EntityParticle WHERE EntityDefinition.QualifiedApiName = '" + this.sobjectName + "'")).then(res => {
+
+      sfConn.rest("/services/data/v" + apiVersion + "/query/?q=" + encodeURIComponent("SELECT QualifiedApiName, EntityDefinitionId, FieldDefinitionId, NamespacePrefix, DeveloperName, MasterLabel, Label, Length, DataType, ServiceDataTypeId, ValueTypeId, ExtraTypeInfo, IsAutonumber, ByteLength, IsCaseSensitive, IsUnique, IsCreatable, IsUpdatable, IsDefaultedOnCreate, IsWriteRequiresMasterRead, IsCalculated, IsHighScaleNumber, IsHtmlFormatted, IsNameField, IsNillable, IsPermissionable, IsEncrypted, Digits, InlineHelpText, RelationshipName, ReferenceTargetField, Name, Mask, MaskType, IsWorkflowFilterable, IsCompactLayoutable, Precision, Scale, IsFieldHistoryTracked, IsApiFilterable, IsApiSortable, IsApiGroupable, IsListVisible, IsLayoutable, IsDependentPicklist, IsDeprecatedAndHidden, IsDisplayLocationInDecimal, DefaultValueFormula, IsIdLookup, IsNamePointing, RelationshipOrder, ReferenceTo, IsComponent, IsCompound,  FieldDefinition.DurableId, EntityDefinition.DurableId FROM EntityParticle WHERE EntityDefinition.QualifiedApiName ='" + this.sobjectName + "'")).then(res => {
         for (let entityParticle of res.records) {
           this.fieldRows.getRow(entityParticle.QualifiedApiName).entityParticle = entityParticle;
           if (!this.entityDefinitionDurableId){